compiles and works on darwin-aarch64
